+++ apps/sv/frontend/src/components/votes/VoteRequest.tsx
@@ -25,7 +25,6 @@ import GrantFeaturedAppRight from './actions/GrantFeaturedAppRight';
 import RemoveFutureCoinConfigSchedule from './actions/RemoveFutureCoinConfigSchedule';
 import RemoveMember from './actions/RemoveMember';
 import RevokeFeaturedAppRight from './actions/RevokeFeaturedAppRight';
-import SetCoinRulesConfig from './actions/SetCoinRulesConfig';
 import SetCoinRulesEnabledChoices from './actions/SetCoinRulesEnabledChoices';
 import SetSvcRulesConfig from './actions/SetSvcRulesConfig';
 import UpdateFutureCoinConfigSchedule from './actions/UpdateFutureCoinConfigSchedule';
@@ -41,7 +40,6 @@ const VoteRequest: React.FC = () => {
     { name: 'Feature Application', value: 'SRARC_GrantFeaturedAppRight' },
     { name: 'Unfeature Application', value: 'SRARC_RevokeFeaturedAppRight' },
     { name: 'Set SvcRules Configuration', value: 'SRARC_SetConfig' },
-    { name: 'Set CoinRules Configuration', value: 'CRARC_SetConfigSchedule' },
     { name: 'Set CoinRules Enabled Choices', value: 'CRARC_SetEnabledChoices' },
     { name: 'Add Coin Configuration Schedule', value: 'CRARC_AddFutureCoinConfigSchedule' },
     { name: 'Remove Coin Configuration Schedule', value: 'CRARC_RemoveFutureCoinConfigSchedule' },
@@ -104,9 +102,6 @@ const VoteRequest: React.FC = () => {
             <RevokeFeaturedAppRight chooseAction={chooseAction} />
           )}
           {actionName === 'SRARC_SetConfig' && <SetSvcRulesConfig chooseAction={chooseAction} />}
-          {actionName === 'CRARC_SetConfigSchedule' && (
-            <SetCoinRulesConfig chooseAction={chooseAction} />
-          )}
           {actionName === 'CRARC_SetEnabledChoices' && (
             <SetCoinRulesEnabledChoices chooseAction={chooseAction} />
           )}